Optimize build scripts and migrate to jewel-laf
