+++ tooling/core/usecase/build.gradle.kts
@@ -1,16 +1,14 @@
 plugins {
-    alias(libs.plugins.kotlinJvm)
+    alias(libs.plugins.intelliJCommon)
     alias(libs.plugins.compose.compiler)
     alias(libs.plugins.jetbrainsCompose)
-}
-
-kotlin {
-    jvmToolchain(17)
 }
 
 dependencies {
     implementation(projects.tooling.core.shared)
     implementation(projects.tooling.core.database)
     implementation(projects.tooling.core.model)
-    implementation(compose.runtime)
+    implementation(compose.runtime) {
+        exclude(group = "org.jetbrains.kotlinx")
+    }
 }
